Hide navigation bar items no longer used on new install
